Rsq intersect moved to utility repo, start on pressure and density implementation
